requirements.txt
asgiref==3.7.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.2
django-allauth==0.63.3
django-modeltranslation==0.19.10
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.2
